app/food_detector_hf.py
"""
Food Detection using Hugging Face Models
"""
import torch
from PIL import Image
from transformers import pipeline, AutoProcessor, AutoModelForObjectDetection
import numpy as np
import cv2
from typing import List, Dict, Tuple
import logging

logger = logging.getLogger(__name__)

class HuggingFaceFoodDetector:
    def __init__(self):
        """Initialize food detection models"""
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        
        # Try different models for food detection
        self.models = self._initialize_models()
        
    def _initialize_models(self):
        """Initialize various models for food detection"""
        models = {}
        
        try:
            # Object detection model (DETR)
            models['object_detector'] = pipeline(
                "object-detection", 
                model="facebook/detr-resnet-50",
                device=0 if torch.cuda.is_available() else -1
            )
        except Exception as e:
            logger.warning(f"Failed to load object detection model: {e}")
            
        try:
            # Image classification for food
            models['food_classifier'] = pipeline(
                "image-classification",
                model="nateraw/food",
                device=0 if torch.cuda.is_available() else -1
            )
        except Exception as e:
            logger.warning(f"Failed to load food classification model: {e}")
            
        try:
            # Vision-language model for more flexible detection
            models['clip_classifier'] = pipeline(
                "zero-shot-image-classification",
                model="openai/clip-vit-base-patch32",
                device=0 if torch.cuda.is_available() else -1
            )
        except Exception as e:
            logger.warning(f"Failed to load CLIP model: {e}")
            
        return models
    
    def detect_food_items(self, image_path: str) -> List[str]:
        """Detect food items in an image using multiple models"""
        detected_items = set()
        
        try:
            # Load image
            image = Image.open(image_path)
            
            # Method 1: Food classifier
            if 'food_classifier' in self.models:
                try:
                    results = self.models['food_classifier'](image, top_k=5)
                    for result in results:
                        if result['score'] > 0.1:
                            # Extract food name from label (often in format "food_name" or "category: food_name")
                            label = result['label'].lower()
                            if ':' in label:
                                label = label.split(':')[-1].strip()
                            label = label.replace('_', ' ').replace('-', ' ')
                            detected_items.add(label)
                except Exception as e:
                    logger.warning(f"Food classifier failed: {e}")
            
            # Method 2: Zero-shot classification with food items
            if 'clip_classifier' in self.models:
                try:
                    # Common food items to check
                    candidate_labels = [
                        "chicken", "beef", "pork", "fish", "shrimp", "egg", "cheese",
                        "tomato", "carrot", "potato", "onion", "garlic", "pepper", "lettuce",
                        "broccoli", "spinach", "cucumber", "corn", "mushroom", "peas",
                        "apple", "banana", "orange", "strawberry", "grapes",
                        "rice", "pasta", "bread", "noodles",
                        "milk", "yogurt", "butter"
                    ]
                    
                    results = self.models['clip_classifier'](
                        image, 
                        candidate_labels=candidate_labels,
                        hypothesis_template="This is a photo of {}"
                    )
                    
                    # Add items with reasonable confidence
                    for i, label in enumerate(results['labels'][:10]):
                        if results['scores'][i] > 0.15:
                            detected_items.add(label)
                except Exception as e:
                    logger.warning(f"CLIP classifier failed: {e}")
            
            # Method 3: Object detection
            if 'object_detector' in self.models:
                try:
                    results = self.models['object_detector'](image)
                    food_related_objects = [
                        'banana', 'apple', 'sandwich', 'orange', 'broccoli', 'carrot',
                        'hot dog', 'pizza', 'donut', 'cake', 'fruit', 'vegetable'
                    ]
                    
                    for detection in results:
                        label = detection['label'].lower()
                        if detection['score'] > 0.5 and any(food in label for food in food_related_objects):
                            detected_items.add(label)
                except Exception as e:
                    logger.warning(f"Object detector failed: {e}")
            
            # Fallback: Image segmentation based detection
            if len(detected_items) < 3:
                detected_items.update(self._segment_based_detection(image))
            
        except Exception as e:
            logger.error(f"Error in food detection: {e}")
            # Return default items
            detected_items = {'chicken', 'tomato', 'onion'}
        
        # Ensure we have at least 3 items
        if len(detected_items) < 3:
            common_foods = ['potato', 'egg', 'rice', 'pasta', 'cheese']
            for food in common_foods:
                if food not in detected_items:
                    detected_items.add(food)
                if len(detected_items) >= 3:
                    break
        
        return list(detected_items)
    
    def _segment_based_detection(self, image: Image) -> List[str]:
        """Segment image and detect food based on regions"""
        detected = []
        
        try:
            # Convert PIL to numpy array
            img_array = np.array(image)
            
            # Simple color-based segmentation
            hsv = cv2.cvtColor(img_array, cv2.COLOR_RGB2HSV)
            
            # Define color ranges for common food items
            color_ranges = {
                'tomato': [(0, 50, 50), (10, 255, 255)],  # Red
                'carrot': [(10, 50, 50), (25, 255, 255)],  # Orange
                'lettuce': [(35, 50, 50), (85, 255, 255)],  # Green
                'potato': [(20, 20, 50), (30, 100, 200)],  # Brown/beige
                'egg': [(20, 10, 200), (30, 30, 255)],  # White/yellow
                'meat': [(0, 20, 50), (20, 100, 150)],  # Dark red/brown
            }
            
            for food, (lower, upper) in color_ranges.items():
                mask = cv2.inRange(hsv, np.array(lower), np.array(upper))
                if cv2.countNonZero(mask) > img_array.shape[0] * img_array.shape[1] * 0.01:  # At least 1% of image
                    detected.append(food)
            
        except Exception as e:
            logger.warning(f"Segment-based detection failed: {e}")
        
        return detected

# Singleton instance
_food_detector = None

def get_food_detector():
    """Get or create the food detector instance"""
    global _food_detector
    if _food_detector is None:
        _food_detector = HuggingFaceFoodDetector()
    return _food_detector