requirements.txt
Flask==3.1.1
flask-login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
email-validator==2.2.0
psycopg2-binary==2.9.10
opencv-python==4.11.0.86
numpy==2.2.3
pandas==2.2.3
scikit-learn==1.6.1
google-api-python-client==2.175.0
bcrypt==4.3.0
gunicorn==23.0.0
Pillow==11.1.0
WTForms==3.2.1
